organize global websocket hooks
