+++ packages/client/src/components/chat/hooks/use-chat-state.tsx
@@ -9,7 +9,7 @@ import { useChatModelControl } from './use-chat-model-control';
 import { ChatMessage } from 'shared/schema';
 import { APIProviders } from 'shared/index';
 import { useUpdateActiveChatTab } from '@/components/global-state/global-helper-hooks';
-import { useActiveChatTab } from '@/components/global-state/global-websocket-selectors';
+import { useActiveChatTab } from '@/components/global-state/websocket-selector-hoooks';
 
 type TempChatMessage = ChatMessage & { tempId?: string };
