+++ packages/client/src/components/global-state/global-helper-hooks.ts
@@ -1,6 +1,6 @@
 import { useCallback } from "react";
 import { v4 as uuidv4 } from "uuid";
-import { useGlobalStateContext } from "@/components/global-state/global-state-websocket";
+import { useGlobalStateContext } from "@/components/global-state/global-state-websocket-handler-context";
 import { useGlobalState } from "./use-global-state";
 import { buildTicketContent } from "@/components/tickets/utils/ticket-utils";
 import type { TicketWithTasks } from "@/hooks/api/use-tickets-api";
@@ -11,7 +11,7 @@ import {
     linkSettingsSchema,
     LinkSettings,
 } from "shared";
-import { useActiveChatTab, useActiveProjectTab } from "./global-websocket-selectors";
+import { useActiveChatTab, useActiveProjectTab } from "./websocket-selector-hoooks";
 
 /**
  * Helper for partial updates of state.
