+++ packages/client/src/components/settings/settings-dialog.tsx
@@ -12,7 +12,7 @@ import {
 import { useUpdateGlobalStateKey } from "@/components/global-state/global-helper-hooks"
 import { Label } from "@/components/ui/label"
 import { Input } from "@/components/ui/input"
-import { useSettings } from "../global-state/global-websocket-selectors"
+import { useSettings } from "../global-state/websocket-selector-hoooks"
 
 type ThemeOption = {
     label: string;
