+++ packages/client/src/routes/projects.tsx
@@ -14,10 +14,10 @@ import { useHotkeys } from 'react-hotkeys-hook'
 import { ProjectsTabManager } from '@/components/tab-managers/projects-tab-manager'
 import { Button } from '@/components/ui/button'
 import { useSelectedFiles } from '@/hooks/utility-hooks/use-selected-files'
-import { useCreateProjectTab, useGlobalStateCore, useUpdateActiveProjectTabStateKey } from '@/components/global-state/global-helper-hooks'
+import { useCreateProjectTab, useUpdateActiveProjectTabStateKey } from '@/components/global-state/global-helper-hooks'
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription, DialogFooter } from '@/components/ui/dialog'
 import { InfoTooltip } from '@/components/info-tooltip'
-import { useActiveProjectTab, useProjectTabs } from '@/components/global-state/global-websocket-selectors'
+import { useActiveProjectTab, useAllProjectTabs } from '@/components/global-state/websocket-selector-hoooks'
 
 export const Route = createFileRoute('/projects')({
     component: ProjectsPage,
@@ -27,7 +27,7 @@ function ProjectsPage() {
     const filePanelRef = useRef<FilePanelRef>(null)
     const promptPanelRef = useRef<PromptOverviewPanelRef>(null)
 
-    const tabs = useProjectTabs()
+    const tabs = useAllProjectTabs()
 
     // new way pulling from the query cahce for the project tabs
     const { tabData: activeTabState, id: projectActiveTabId } = useActiveProjectTab()
